// codegapp/src/app/Models/Pessoa.ts
export class Pessoa {

    // constructor() {
    //     this.id = 0;
    //     this.nome = "";
    //     // this.dataNascimento = Date.now;
    //     this.cpf = "";
    //     // this.dataCliente
    // }

    pessoaId: number = 0;
    nome?: string;
    dataNascimento?: Date;
    cpf?: string;
    isFuncionario?: boolean;
    projetoId?: number;
}