// codegapp/src/app/Models/Projeto.ts
export class Projeto {

    projetoId: number = 0;
    nome?: string;
    dataInicio?: Date;
    dataPrevisaoFim?: Date;
    dataFim?: Date;
    descricao?: string;
    status?: string;
    orcamento: number = 0;
    risco?: string;
    pessoaId: number = 0;
    // Pessoa: Pessoa;
    // Membro: Membro;

}